functions/requirements.txt
# Core ML Libraries
tensorflow==2.16.1
tensorflow-decision-forests==1.9.0

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

# Data Processing
numpy==1.26.2
pandas==2.1.3

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Firebase Integration
firebase-admin==6.3.0
firebase-functions==0.4.0

# HTTP Requests
requests==2.31.0

# Cloud Functions Framework
functions-framework==3.5.0

pytz